correção nos arquivos de requisições
