+++ src/apis/authApi.ts
@@ -9,3 +9,20 @@ export const login = async (formLoginData: ILogin) => {
     throw error.response?.data || error;
   }
 };
+
+export const logout = async () => {
+  try {
+    const response = await api.post(
+      "/logout",
+      {},
+      {
+        headers: {
+          Authorization: `${localStorage.getItem("token")}`,
+        },
+      }
+    );
+    return response.data;
+  } catch (error: any) {
+    throw error.response?.data || error;
+  }
+};
